Fix 'filter by date' in admin orders page.
The bug occurred when you did something similar to the following:
* Click "show orders delivered today".
* Click "clear all filters".
* Click "show orders delivered today" again.

If there were orders listed the first time the filter was applied,
those orders will not appear the second time the filter is applied.

This commit fixes this by updating to the version of ng-grid's
`Grid.clearAllFilters`, which got merged into upstream.